refactor(backend): split A2A schedule service domains (#458)
